Resolve Compiler Warnings (#64)

+++ Source/Core/Editor/Widgets/GUI_Widget_FramerateGraph.cpp
@@ -13,7 +13,7 @@ void Widget_FramerateGraph::Draw() {
 
 
     // Cap Array Length At Target Set
-    if (FramerateHistory_.size() > HistoryLength_) {
+    if ((long)FramerateHistory_.size() > (long)HistoryLength_) {
         FramerateHistory_.erase(FramerateHistory_.begin());
     } else {
